funcionalidade editar e selecionar animais

+++ js/fichaCrtl.js
@@ -4,22 +4,25 @@ angular.module("ficha").controller('fichaCrtl', ['$scope', 'fichaService', 'anim
         $scope.mostrarBuscaDeFichas = true;
         $scope.mostrarCadastraFicha = false;
         $scope.mostrarFichasCadastradas = false;
-
         $scope.excluir = function (ficha) {
-            fichaService.remove(ficha);
-            $scope.fichas = fichaService.getAll();
+            if (confirm('Tem certeza?')){
+                fichaService.remove(ficha);
+                $scope.fichas = fichaService.getAll();
+            }
         }
 
         $scope.showCadastro = function () {
+            $scope.animais = animalService.getAll();
             $scope.mostrarBuscaDeFichas = false;
             $scope.mostrarCadastraFicha = true;
             $scope.mostrarFichasCadastradas = false;
-            $scope.animais = animalService.getAll();
+            $scope.ficha = { status: 'Ativo', dataDeCadastro: new Date() };
         }
 
         $scope.cancelarCadastro = function () {
             $scope.mostrarCadastraFicha = false;
             $scope.mostrarBuscaDeFichas = true;
+            delete $scope.ficha;
         }
 
         $scope.buscar = function () {
@@ -27,8 +30,20 @@ angular.module("ficha").controller('fichaCrtl', ['$scope', 'fichaService', 'anim
             $scope.mostrarFichasCadastradas = true;
         }
 
-        $scope.cadastrarFicha = function (ficha) {
-            fichaService.add(ficha);
+        $scope.editar = function(ficha){
+            $scope.ficha = angular.copy(ficha);
+            $scope.ficha.dataDeCadastro = new Date($scope.ficha.dataDeCadastro);
+            $scope.mostrarBuscaDeFichas = false;
+            $scope.mostrarCadastraFicha = true;
+            $scope.mostrarFichasCadastradas = false;
+        }
+
+        $scope.salvar = function (ficha) {
+            if (ficha.id) {
+                fichaService.update(angular.copy(ficha));
+            } else {
+                fichaService.add(ficha);
+            }
             $scope.fichas = fichaService.getAll();
             $scope.mostrarBuscaDeFichas = true;
             $scope.mostrarCadastraFicha = false;
